server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
google-auth==2.35.0
requests==2.32.3
PyJWT==2.9.0
langchain_community>=0.3.15
langchain-core>=0.3.68
